Moved css file to stylesheets folder, updated import to be correct
